project/requirements.txt
argon2-cffi==19.2.0
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.2
Django==2.2.5
django-crispy-forms==1.8.0
Pillow==6.2.1
pycparser==2.19
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
wincertstore==0.2
